Рефакторинг

+++ src/include/utils.h
@@ -10,15 +10,15 @@ namespace utils
 {
     std::string escapeString(const std::string& str);
     std::string getLineFromString(const std::string& str, int line);
-    int linenoFromPosition(const std::string& str, size_t position);
+    size_t linenoFromPosition(const std::string& str, size_t position);
 
     // Визначає позицію в стрічці із загальної позиції символа в стрічці
     size_t positionInLineFromPosition(const std::string& str, size_t position);
     void replaceTabToSpace(std::string& str);
-    
+
     // Повертає кількість символів в utf8 стрічці
     size_t utf8Size(const std::string& str);
-    
+
     // Відступ на задану ширину, заповнює відступ пробілами.
     // Наприклад:
     //     std::cout << "id =" << indent(10) << "10";
